+++ Cargo.toml
@@ -12,7 +12,7 @@ categories = ["embedded", "hardware-support", "no-std"]
 readme = "README.md"
 
 [dependencies]
-embedded-hal = { version = "1.0.0" }
+embedded-hal = { version = "1.0.0", optional = true }
 heapless = "0.8.0"
 embedded-graphics = { version = "0.8.1", default-features = false }
 embedded-graphics-core = "0.4.0"
@@ -20,5 +20,6 @@ embedded-graphics-core = "0.4.0"
 embedded-hal-async = { version = "1.0.0", optional = true }
 
 [features]
+sync = ["dep:embedded-hal"]
 async = ["dep:embedded-hal-async"]
 std = []
